[mod] #29 JwtTokenResponse를 실제 응답 response 형태로 변경
