--- src/main/java/org/hilingual/domain/security/token/api/dto/res/JwtTokenResponse.java
@@ -1,19 +1,0 @@
-package org.hilingual.domain.security.token.api.dto.res;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class JwtTokenResponse {
-    private String accessToken;
-    private String refreshToken;
-
-    public static JwtTokenResponse of(String accessToken, String refreshToken) {
-        return new JwtTokenResponse(accessToken, refreshToken);
-    }
-}
